Factor most of prowgen in a non-main package
