+++ cmd/ci-operator-prowgen/main_test.go
@@ -4,7 +4,6 @@ import (
 	"flag"
 	"fmt"
 	"io/ioutil"
-	"log"
 	"os"
 	"path/filepath"
 	"reflect"
@@ -14,13 +13,8 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/pmezard/go-difflib/difflib"
-	kubeapi "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/equality"
-	"k8s.io/apimachinery/pkg/api/resource"
-	v1 "k8s.io/test-infra/prow/apis/prowjobs/v1"
 	prowconfig "k8s.io/test-infra/prow/config"
 
-	ciop "github.com/openshift/ci-tools/pkg/api"
 	"github.com/openshift/ci-tools/pkg/config"
 	"github.com/openshift/ci-tools/pkg/jobconfig"
 )
@@ -32,1493 +26,6 @@ var unexportedFields = []cmp.Option{
 	cmpopts.IgnoreUnexported(prowconfig.Periodic{}),
 	cmpopts.IgnoreUnexported(prowconfig.Brancher{}),
 	cmpopts.IgnoreUnexported(prowconfig.RegexpChangeMatcher{}),
-}
-
-func TestGeneratePodSpec(t *testing.T) {
-	tests := []struct {
-		description    string
-		info           *prowgenInfo
-		secrets        []*ciop.Secret
-		targets        []string
-		additionalArgs []string
-
-		expected *kubeapi.PodSpec
-	}{
-		{
-			description: "standard use case",
-			info:        &prowgenInfo{Info: config.Info{Org: "org", Repo: "repo", Branch: "branch"}},
-			secrets:     nil,
-			targets:     []string{"target"},
-
-			expected: &kubeapi.PodSpec{
-				ServiceAccountName: "ci-operator",
-				Containers: []kubeapi.Container{{
-					Image:           "ci-operator:latest",
-					ImagePullPolicy: kubeapi.PullAlways,
-					Command:         []string{"ci-operator"},
-					Args: []string{
-						"--give-pr-author-access-to-namespace=true",
-						"--artifact-dir=$(ARTIFACTS)",
-						"--sentry-dsn-path=/etc/sentry-dsn/ci-operator",
-						"--kubeconfig=/etc/apici/kubeconfig",
-						"--image-import-pull-secret=/etc/pull-secret/.dockerconfigjson",
-						"--target=target",
-					},
-					Resources: kubeapi.ResourceRequirements{
-						Requests: kubeapi.ResourceList{"cpu": *resource.NewMilliQuantity(10, resource.DecimalSI)},
-					},
-					VolumeMounts: []kubeapi.VolumeMount{{Name: "sentry-dsn", MountPath: "/etc/sentry-dsn", ReadOnly: true},
-						{Name: "apici-ci-operator-credentials", ReadOnly: true, MountPath: "/etc/apici"},
-						{Name: "pull-secret", ReadOnly: true, MountPath: "/etc/pull-secret"}},
-				}},
-				Volumes: []kubeapi.Volume{{
-					Name: "sentry-dsn",
-					VolumeSource: kubeapi.VolumeSource{
-						Secret: &kubeapi.SecretVolumeSource{SecretName: "sentry-dsn"},
-					},
-				},
-					{
-						Name: "apici-ci-operator-credentials",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "apici-ci-operator-credentials", Items: []kubeapi.KeyToPath{{Key: "sa.ci-operator.apici.config", Path: "kubeconfig"}}},
-						},
-					},
-					{
-						Name: "pull-secret",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "regcred"},
-						},
-					},
-				},
-			},
-		},
-		{
-			description:    "additional args are included in podspec",
-			info:           &prowgenInfo{Info: config.Info{Org: "org", Repo: "repo", Branch: "branch"}},
-			secrets:        nil,
-			targets:        []string{"target"},
-			additionalArgs: []string{"--promote", "--some=thing"},
-
-			expected: &kubeapi.PodSpec{
-				ServiceAccountName: "ci-operator",
-				Containers: []kubeapi.Container{{
-					Image:           "ci-operator:latest",
-					ImagePullPolicy: kubeapi.PullAlways,
-					Command:         []string{"ci-operator"},
-					Args: []string{
-						"--give-pr-author-access-to-namespace=true",
-						"--artifact-dir=$(ARTIFACTS)",
-						"--sentry-dsn-path=/etc/sentry-dsn/ci-operator",
-						"--kubeconfig=/etc/apici/kubeconfig",
-						"--image-import-pull-secret=/etc/pull-secret/.dockerconfigjson",
-						"--promote",
-						"--some=thing",
-						"--target=target",
-					},
-					Resources: kubeapi.ResourceRequirements{
-						Requests: kubeapi.ResourceList{"cpu": *resource.NewMilliQuantity(10, resource.DecimalSI)},
-					},
-					VolumeMounts: []kubeapi.VolumeMount{{Name: "sentry-dsn", MountPath: "/etc/sentry-dsn", ReadOnly: true},
-						{Name: "apici-ci-operator-credentials", ReadOnly: true, MountPath: "/etc/apici"},
-						{Name: "pull-secret", ReadOnly: true, MountPath: "/etc/pull-secret"}},
-				}},
-				Volumes: []kubeapi.Volume{{
-					Name: "sentry-dsn",
-					VolumeSource: kubeapi.VolumeSource{
-						Secret: &kubeapi.SecretVolumeSource{SecretName: "sentry-dsn"},
-					},
-				},
-					{
-						Name: "apici-ci-operator-credentials",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "apici-ci-operator-credentials", Items: []kubeapi.KeyToPath{{Key: "sa.ci-operator.apici.config", Path: "kubeconfig"}}},
-						},
-					},
-					{
-						Name: "pull-secret",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "regcred"},
-						},
-					},
-				},
-			},
-		},
-		{
-			description:    "additional args and secret are included in podspec",
-			info:           &prowgenInfo{Info: config.Info{Org: "org", Repo: "repo", Branch: "branch"}},
-			secrets:        []*ciop.Secret{{Name: "secret-name", MountPath: "/usr/local/test-secret"}},
-			targets:        []string{"target"},
-			additionalArgs: []string{"--promote", "--some=thing"},
-
-			expected: &kubeapi.PodSpec{
-				ServiceAccountName: "ci-operator",
-				Containers: []kubeapi.Container{{
-					Image:           "ci-operator:latest",
-					ImagePullPolicy: kubeapi.PullAlways,
-					Command:         []string{"ci-operator"},
-					Args: []string{
-						"--give-pr-author-access-to-namespace=true",
-						"--artifact-dir=$(ARTIFACTS)",
-						"--sentry-dsn-path=/etc/sentry-dsn/ci-operator",
-						"--kubeconfig=/etc/apici/kubeconfig",
-						"--image-import-pull-secret=/etc/pull-secret/.dockerconfigjson",
-						"--promote",
-						"--some=thing",
-						"--target=target",
-						"--secret-dir=/secrets/secret-name",
-					},
-					Resources: kubeapi.ResourceRequirements{
-						Requests: kubeapi.ResourceList{"cpu": *resource.NewMilliQuantity(10, resource.DecimalSI)},
-					},
-					VolumeMounts: []kubeapi.VolumeMount{
-						{Name: "sentry-dsn", MountPath: "/etc/sentry-dsn", ReadOnly: true},
-						{Name: "apici-ci-operator-credentials", ReadOnly: true, MountPath: "/etc/apici"},
-						{Name: "pull-secret", ReadOnly: true, MountPath: "/etc/pull-secret"},
-						{Name: "secret-name", MountPath: "/secrets/secret-name", ReadOnly: true},
-					},
-				}},
-				Volumes: []kubeapi.Volume{
-					{
-						Name: "sentry-dsn",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "sentry-dsn"},
-						},
-					},
-					{
-						Name: "apici-ci-operator-credentials",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "apici-ci-operator-credentials", Items: []kubeapi.KeyToPath{{Key: "sa.ci-operator.apici.config", Path: "kubeconfig"}}},
-						},
-					},
-					{
-						Name: "pull-secret",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "regcred"},
-						},
-					},
-					{
-						Name: "secret-name",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "secret-name"},
-						},
-					},
-				},
-			},
-		},
-		{
-			description: "multiple targets",
-			info:        &prowgenInfo{Info: config.Info{Org: "org", Repo: "repo", Branch: "branch"}},
-			secrets:     nil,
-			targets:     []string{"target", "more", "and-more"},
-
-			expected: &kubeapi.PodSpec{
-				ServiceAccountName: "ci-operator",
-				Containers: []kubeapi.Container{{
-					Image:           "ci-operator:latest",
-					ImagePullPolicy: kubeapi.PullAlways,
-					Command:         []string{"ci-operator"},
-					Args: []string{
-						"--give-pr-author-access-to-namespace=true",
-						"--artifact-dir=$(ARTIFACTS)",
-						"--sentry-dsn-path=/etc/sentry-dsn/ci-operator",
-						"--kubeconfig=/etc/apici/kubeconfig",
-						"--image-import-pull-secret=/etc/pull-secret/.dockerconfigjson",
-						"--target=target",
-						"--target=more",
-						"--target=and-more",
-					},
-					Resources: kubeapi.ResourceRequirements{
-						Requests: kubeapi.ResourceList{"cpu": *resource.NewMilliQuantity(10, resource.DecimalSI)},
-					},
-					VolumeMounts: []kubeapi.VolumeMount{{Name: "sentry-dsn", MountPath: "/etc/sentry-dsn", ReadOnly: true},
-						{Name: "apici-ci-operator-credentials", ReadOnly: true, MountPath: "/etc/apici"},
-						{Name: "pull-secret", ReadOnly: true, MountPath: "/etc/pull-secret"}},
-				}},
-				Volumes: []kubeapi.Volume{{
-					Name: "sentry-dsn",
-					VolumeSource: kubeapi.VolumeSource{
-						Secret: &kubeapi.SecretVolumeSource{SecretName: "sentry-dsn"},
-					},
-				}, {
-					Name: "apici-ci-operator-credentials",
-					VolumeSource: kubeapi.VolumeSource{
-						Secret: &kubeapi.SecretVolumeSource{SecretName: "apici-ci-operator-credentials", Items: []kubeapi.KeyToPath{{Key: "sa.ci-operator.apici.config", Path: "kubeconfig"}}},
-					},
-				},
-					{
-						Name: "pull-secret",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "regcred"},
-						},
-					}},
-			},
-		},
-		{
-			description: "private job",
-			info: &prowgenInfo{
-				Info:   config.Info{Org: "org", Repo: "repo", Branch: "branch"},
-				config: config.Prowgen{Private: true},
-			},
-			secrets: nil,
-			targets: []string{"target"},
-
-			expected: &kubeapi.PodSpec{
-				ServiceAccountName: "ci-operator",
-				Containers: []kubeapi.Container{
-					{
-						Image:           "ci-operator:latest",
-						ImagePullPolicy: kubeapi.PullAlways,
-						Command:         []string{"ci-operator"},
-						Args: []string{
-							"--give-pr-author-access-to-namespace=true",
-							"--artifact-dir=$(ARTIFACTS)",
-							"--sentry-dsn-path=/etc/sentry-dsn/ci-operator",
-							"--kubeconfig=/etc/apici/kubeconfig",
-							"--image-import-pull-secret=/etc/pull-secret/.dockerconfigjson",
-							"--target=target",
-							"--oauth-token-path=/usr/local/github-credentials/oauth",
-						},
-						Resources: kubeapi.ResourceRequirements{
-							Requests: kubeapi.ResourceList{"cpu": *resource.NewMilliQuantity(10, resource.DecimalSI)},
-						},
-						VolumeMounts: []kubeapi.VolumeMount{
-							{
-								Name: "sentry-dsn", MountPath: "/etc/sentry-dsn", ReadOnly: true,
-							},
-							{Name: "apici-ci-operator-credentials", ReadOnly: true, MountPath: "/etc/apici"},
-							{Name: "pull-secret", ReadOnly: true, MountPath: "/etc/pull-secret"},
-							{
-								Name:      "github-credentials-openshift-ci-robot-private-git-cloner",
-								MountPath: "/usr/local/github-credentials",
-								ReadOnly:  true,
-							},
-						},
-					},
-				},
-				Volumes: []kubeapi.Volume{
-					{
-						Name: "sentry-dsn",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "sentry-dsn"},
-						},
-					},
-					{
-						Name: "apici-ci-operator-credentials",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "apici-ci-operator-credentials", Items: []kubeapi.KeyToPath{{Key: "sa.ci-operator.apici.config", Path: "kubeconfig"}}},
-						},
-					},
-					{
-						Name: "pull-secret",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "regcred"},
-						},
-					},
-					{
-						Name: "github-credentials-openshift-ci-robot-private-git-cloner",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "github-credentials-openshift-ci-robot-private-git-cloner"},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.description, func(t *testing.T) {
-			var podSpec *kubeapi.PodSpec
-			podSpec = generateCiOperatorPodSpec(tc.info, tc.secrets, tc.targets, tc.additionalArgs...)
-			if !equality.Semantic.DeepEqual(podSpec, tc.expected) {
-				t.Errorf("%s: expected PodSpec diff:\n%s", tc.description, cmp.Diff(tc.expected, podSpec, unexportedFields...))
-			}
-		})
-	}
-}
-
-func TestGeneratePodSpecTemplate(t *testing.T) {
-	tests := []struct {
-		info    *prowgenInfo
-		release string
-		test    ciop.TestStepConfiguration
-
-		expected *kubeapi.PodSpec
-	}{
-		{
-			info:    &prowgenInfo{Info: config.Info{Org: "organization", Repo: "repo", Branch: "branch"}},
-			release: "origin-v4.0",
-			test: ciop.TestStepConfiguration{
-				As:       "test",
-				Commands: "commands",
-				OpenshiftAnsibleClusterTestConfiguration: &ciop.OpenshiftAnsibleClusterTestConfiguration{
-					ClusterTestConfiguration: ciop.ClusterTestConfiguration{ClusterProfile: "gcp"},
-				},
-			},
-
-			expected: &kubeapi.PodSpec{
-				ServiceAccountName: "ci-operator",
-				Volumes: []kubeapi.Volume{
-					{
-						Name: "sentry-dsn",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "sentry-dsn"},
-						},
-					},
-					{
-						Name: "apici-ci-operator-credentials",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "apici-ci-operator-credentials", Items: []kubeapi.KeyToPath{{Key: "sa.ci-operator.apici.config", Path: "kubeconfig"}}},
-						},
-					},
-					{
-						Name: "pull-secret",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "regcred"},
-						},
-					},
-					{
-						Name: "job-definition",
-						VolumeSource: kubeapi.VolumeSource{
-							ConfigMap: &kubeapi.ConfigMapVolumeSource{
-								LocalObjectReference: kubeapi.LocalObjectReference{
-									Name: "prow-job-cluster-launch-e2e",
-								},
-							},
-						},
-					},
-					{
-						Name: "cluster-profile",
-						VolumeSource: kubeapi.VolumeSource{
-							Projected: &kubeapi.ProjectedVolumeSource{
-								Sources: []kubeapi.VolumeProjection{
-									{
-										Secret: &kubeapi.SecretProjection{
-											LocalObjectReference: kubeapi.LocalObjectReference{
-												Name: "cluster-secrets-gcp",
-											},
-										},
-									},
-									{
-										ConfigMap: &kubeapi.ConfigMapProjection{
-											LocalObjectReference: kubeapi.LocalObjectReference{
-												Name: "cluster-profile-gcp",
-											},
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-				Containers: []kubeapi.Container{{
-					Image:           "ci-operator:latest",
-					ImagePullPolicy: kubeapi.PullAlways,
-					Command:         []string{"ci-operator"},
-					Args: []string{
-						"--give-pr-author-access-to-namespace=true",
-						"--artifact-dir=$(ARTIFACTS)",
-						"--sentry-dsn-path=/etc/sentry-dsn/ci-operator",
-						"--kubeconfig=/etc/apici/kubeconfig",
-						"--image-import-pull-secret=/etc/pull-secret/.dockerconfigjson",
-						"--target=test",
-						"--secret-dir=/usr/local/test-cluster-profile",
-						"--template=/usr/local/test"},
-					Resources: kubeapi.ResourceRequirements{
-						Requests: kubeapi.ResourceList{"cpu": *resource.NewMilliQuantity(10, resource.DecimalSI)},
-					},
-					Env: []kubeapi.EnvVar{
-						{Name: "CLUSTER_TYPE", Value: "gcp"},
-						{Name: "JOB_NAME_SAFE", Value: "test"},
-						{Name: "TEST_COMMAND", Value: "commands"},
-						{Name: "RPM_REPO_OPENSHIFT_ORIGIN", Value: "https://rpms.svc.ci.openshift.org/openshift-origin-v4.0/"},
-					},
-					VolumeMounts: []kubeapi.VolumeMount{
-						{Name: "sentry-dsn", MountPath: "/etc/sentry-dsn", ReadOnly: true},
-						{Name: "apici-ci-operator-credentials", ReadOnly: true, MountPath: "/etc/apici"},
-						{Name: "pull-secret", ReadOnly: true, MountPath: "/etc/pull-secret"},
-						{Name: "cluster-profile", MountPath: "/usr/local/test-cluster-profile"},
-						{Name: "job-definition", MountPath: "/usr/local/test", SubPath: "cluster-launch-e2e.yaml"},
-					},
-				}},
-			},
-		},
-		{
-			info:    &prowgenInfo{Info: config.Info{Org: "organization", Repo: "repo", Branch: "branch"}},
-			release: "origin-v4.0",
-			test: ciop.TestStepConfiguration{
-				As:       "test",
-				Commands: "commands",
-				OpenshiftInstallerClusterTestConfiguration: &ciop.OpenshiftInstallerClusterTestConfiguration{
-					ClusterTestConfiguration: ciop.ClusterTestConfiguration{ClusterProfile: "aws"},
-				},
-			},
-
-			expected: &kubeapi.PodSpec{
-				ServiceAccountName: "ci-operator",
-				Volumes: []kubeapi.Volume{
-					{
-						Name: "sentry-dsn",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "sentry-dsn"},
-						},
-					},
-					{
-						Name: "apici-ci-operator-credentials",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "apici-ci-operator-credentials", Items: []kubeapi.KeyToPath{{Key: "sa.ci-operator.apici.config", Path: "kubeconfig"}}},
-						},
-					},
-					{
-						Name: "pull-secret",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "regcred"},
-						},
-					},
-					{
-						Name: "job-definition",
-						VolumeSource: kubeapi.VolumeSource{
-							ConfigMap: &kubeapi.ConfigMapVolumeSource{
-								LocalObjectReference: kubeapi.LocalObjectReference{
-									Name: "prow-job-cluster-launch-installer-e2e",
-								},
-							},
-						},
-					},
-					{
-						Name: "cluster-profile",
-						VolumeSource: kubeapi.VolumeSource{
-							Projected: &kubeapi.ProjectedVolumeSource{
-								Sources: []kubeapi.VolumeProjection{
-									{
-										Secret: &kubeapi.SecretProjection{
-											LocalObjectReference: kubeapi.LocalObjectReference{
-												Name: "cluster-secrets-aws",
-											},
-										},
-									},
-								},
-							},
-						},
-					},
-					{
-						Name: "boskos",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "boskos-credentials", Items: []kubeapi.KeyToPath{{Key: "password", Path: "password"}}},
-						},
-					},
-				},
-				Containers: []kubeapi.Container{{
-					Image:           "ci-operator:latest",
-					ImagePullPolicy: kubeapi.PullAlways,
-					Command:         []string{"ci-operator"},
-					Args: []string{
-						"--give-pr-author-access-to-namespace=true",
-						"--artifact-dir=$(ARTIFACTS)",
-						"--sentry-dsn-path=/etc/sentry-dsn/ci-operator",
-						"--kubeconfig=/etc/apici/kubeconfig",
-						"--image-import-pull-secret=/etc/pull-secret/.dockerconfigjson",
-						"--target=test",
-						"--secret-dir=/usr/local/test-cluster-profile",
-						"--template=/usr/local/test",
-						"--lease-server=https://boskos-ci.svc.ci.openshift.org",
-						"--lease-server-username=ci",
-						"--lease-server-password-file=/etc/boskos/password",
-					},
-					Resources: kubeapi.ResourceRequirements{
-						Requests: kubeapi.ResourceList{"cpu": *resource.NewMilliQuantity(10, resource.DecimalSI)},
-					},
-					Env: []kubeapi.EnvVar{
-						{Name: "CLUSTER_TYPE", Value: "aws"},
-						{Name: "JOB_NAME_SAFE", Value: "test"},
-						{Name: "TEST_COMMAND", Value: "commands"},
-					},
-					VolumeMounts: []kubeapi.VolumeMount{
-						{Name: "sentry-dsn", MountPath: "/etc/sentry-dsn", ReadOnly: true},
-						{Name: "apici-ci-operator-credentials", ReadOnly: true, MountPath: "/etc/apici"},
-						{Name: "pull-secret", ReadOnly: true, MountPath: "/etc/pull-secret"},
-						{Name: "boskos", ReadOnly: true, MountPath: "/etc/boskos"},
-						{Name: "cluster-profile", MountPath: "/usr/local/test-cluster-profile"},
-						{Name: "job-definition", MountPath: "/usr/local/test", SubPath: "cluster-launch-installer-e2e.yaml"},
-					},
-				}},
-			},
-		},
-		{
-			info:    &prowgenInfo{Info: config.Info{Org: "organization", Repo: "repo", Branch: "branch"}},
-			release: "origin-v4.0",
-			test: ciop.TestStepConfiguration{
-				As: "test",
-				MultiStageTestConfiguration: &ciop.MultiStageTestConfiguration{
-					ClusterProfile: ciop.ClusterProfileAWS,
-				},
-			},
-			expected: &kubeapi.PodSpec{
-				ServiceAccountName: "ci-operator",
-				Volumes: []kubeapi.Volume{
-					{
-						Name: "sentry-dsn",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "sentry-dsn"},
-						},
-					},
-					{
-						Name: "apici-ci-operator-credentials",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "apici-ci-operator-credentials", Items: []kubeapi.KeyToPath{{Key: "sa.ci-operator.apici.config", Path: "kubeconfig"}}},
-						},
-					},
-					{
-						Name: "pull-secret",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "regcred"},
-						},
-					},
-					{
-						Name: "cluster-profile",
-						VolumeSource: kubeapi.VolumeSource{
-							Projected: &kubeapi.ProjectedVolumeSource{
-								Sources: []kubeapi.VolumeProjection{
-									{
-										Secret: &kubeapi.SecretProjection{
-											LocalObjectReference: kubeapi.LocalObjectReference{
-												Name: "cluster-secrets-aws",
-											},
-										},
-									},
-								},
-							},
-						},
-					},
-					{
-						Name: "boskos",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "boskos-credentials", Items: []kubeapi.KeyToPath{{Key: "password", Path: "password"}}},
-						},
-					},
-				},
-				Containers: []kubeapi.Container{{
-					Image:           "ci-operator:latest",
-					ImagePullPolicy: kubeapi.PullAlways,
-					Command:         []string{"ci-operator"},
-					Args: []string{
-						"--give-pr-author-access-to-namespace=true",
-						"--artifact-dir=$(ARTIFACTS)",
-						"--sentry-dsn-path=/etc/sentry-dsn/ci-operator",
-						"--kubeconfig=/etc/apici/kubeconfig",
-						"--image-import-pull-secret=/etc/pull-secret/.dockerconfigjson",
-						"--target=test",
-						"--secret-dir=/usr/local/test-cluster-profile",
-						"--lease-server=https://boskos-ci.svc.ci.openshift.org",
-						"--lease-server-username=ci",
-						"--lease-server-password-file=/etc/boskos/password",
-					},
-					Resources: kubeapi.ResourceRequirements{
-						Requests: kubeapi.ResourceList{"cpu": *resource.NewMilliQuantity(10, resource.DecimalSI)},
-					},
-					VolumeMounts: []kubeapi.VolumeMount{
-						{Name: "sentry-dsn", MountPath: "/etc/sentry-dsn", ReadOnly: true},
-						{Name: "apici-ci-operator-credentials", ReadOnly: true, MountPath: "/etc/apici"},
-						{Name: "pull-secret", ReadOnly: true, MountPath: "/etc/pull-secret"},
-						{Name: "boskos", ReadOnly: true, MountPath: "/etc/boskos"},
-						{Name: "cluster-profile", MountPath: "/usr/local/test-cluster-profile"},
-					},
-				}},
-			},
-		},
-		{
-			info:    &prowgenInfo{Info: config.Info{Org: "organization", Repo: "repo", Branch: "branch"}},
-			release: "origin-v4.0",
-			test: ciop.TestStepConfiguration{
-				As:       "test",
-				Commands: "commands",
-				OpenshiftInstallerCustomTestImageClusterTestConfiguration: &ciop.OpenshiftInstallerCustomTestImageClusterTestConfiguration{
-					ClusterTestConfiguration: ciop.ClusterTestConfiguration{ClusterProfile: "gcp"},
-					From:                     "pipeline:kubevirt-test",
-					EnableNestedVirt:         true,
-					NestedVirtImage:          "nested-virt-image-name",
-				},
-			},
-
-			expected: &kubeapi.PodSpec{
-				ServiceAccountName: "ci-operator",
-				Volumes: []kubeapi.Volume{
-					{
-						Name: "sentry-dsn",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "sentry-dsn"},
-						},
-					},
-					{
-						Name: "apici-ci-operator-credentials",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "apici-ci-operator-credentials", Items: []kubeapi.KeyToPath{{Key: "sa.ci-operator.apici.config", Path: "kubeconfig"}}},
-						},
-					},
-					{
-						Name: "pull-secret",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "regcred"},
-						},
-					},
-					{
-						Name: "job-definition",
-						VolumeSource: kubeapi.VolumeSource{
-							ConfigMap: &kubeapi.ConfigMapVolumeSource{
-								LocalObjectReference: kubeapi.LocalObjectReference{
-									Name: "prow-job-cluster-launch-installer-custom-test-image",
-								},
-							},
-						},
-					},
-					{
-						Name: "cluster-profile",
-						VolumeSource: kubeapi.VolumeSource{
-							Projected: &kubeapi.ProjectedVolumeSource{
-								Sources: []kubeapi.VolumeProjection{
-									{
-										Secret: &kubeapi.SecretProjection{
-											LocalObjectReference: kubeapi.LocalObjectReference{
-												Name: "cluster-secrets-gcp",
-											},
-										},
-									},
-									{
-										ConfigMap: &kubeapi.ConfigMapProjection{
-											LocalObjectReference: kubeapi.LocalObjectReference{
-												Name: "cluster-profile-gcp",
-											},
-										},
-									},
-								},
-							},
-						},
-					},
-					{
-						Name: "boskos",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "boskos-credentials", Items: []kubeapi.KeyToPath{{Key: "password", Path: "password"}}},
-						},
-					},
-				},
-				Containers: []kubeapi.Container{{
-					Image:           "ci-operator:latest",
-					ImagePullPolicy: kubeapi.PullAlways,
-					Command:         []string{"ci-operator"},
-					Args: []string{
-						"--give-pr-author-access-to-namespace=true",
-						"--artifact-dir=$(ARTIFACTS)",
-						"--sentry-dsn-path=/etc/sentry-dsn/ci-operator",
-						"--kubeconfig=/etc/apici/kubeconfig",
-						"--image-import-pull-secret=/etc/pull-secret/.dockerconfigjson",
-						"--target=test",
-						"--secret-dir=/usr/local/test-cluster-profile",
-						"--template=/usr/local/test",
-						"--lease-server=https://boskos-ci.svc.ci.openshift.org",
-						"--lease-server-username=ci",
-						"--lease-server-password-file=/etc/boskos/password",
-					},
-					Resources: kubeapi.ResourceRequirements{
-						Requests: kubeapi.ResourceList{"cpu": *resource.NewMilliQuantity(10, resource.DecimalSI)},
-					},
-					Env: []kubeapi.EnvVar{
-						{Name: "CLUSTER_TYPE", Value: "gcp"},
-						{Name: "JOB_NAME_SAFE", Value: "test"},
-						{Name: "TEST_COMMAND", Value: "commands"},
-						{Name: "TEST_IMAGESTREAM_TAG", Value: "pipeline:kubevirt-test"},
-						{Name: "CLUSTER_ENABLE_NESTED_VIRT", Value: "true"},
-						{Name: "CLUSTER_NESTED_VIRT_IMAGE", Value: "nested-virt-image-name"},
-					},
-					VolumeMounts: []kubeapi.VolumeMount{
-						{Name: "sentry-dsn", MountPath: "/etc/sentry-dsn", ReadOnly: true},
-						{Name: "apici-ci-operator-credentials", ReadOnly: true, MountPath: "/etc/apici"},
-						{Name: "pull-secret", ReadOnly: true, MountPath: "/etc/pull-secret"},
-						{Name: "boskos", ReadOnly: true, MountPath: "/etc/boskos"},
-						{Name: "cluster-profile", MountPath: "/usr/local/test-cluster-profile"},
-						{Name: "job-definition", MountPath: "/usr/local/test", SubPath: "cluster-launch-installer-custom-test-image.yaml"},
-					},
-				}},
-			},
-		},
-		{
-			info:    &prowgenInfo{Info: config.Info{Org: "organization", Repo: "repo", Branch: "branch"}},
-			release: "origin-v4.0",
-			test: ciop.TestStepConfiguration{
-				As:       "test",
-				Commands: "commands",
-				OpenshiftInstallerCustomTestImageClusterTestConfiguration: &ciop.OpenshiftInstallerCustomTestImageClusterTestConfiguration{
-					ClusterTestConfiguration: ciop.ClusterTestConfiguration{ClusterProfile: "gcp"},
-					From:                     "pipeline:kubevirt-test",
-					EnableNestedVirt:         true,
-				},
-			},
-
-			expected: &kubeapi.PodSpec{
-				ServiceAccountName: "ci-operator",
-				Volumes: []kubeapi.Volume{
-					{
-						Name: "sentry-dsn",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "sentry-dsn"},
-						},
-					},
-					{
-						Name: "apici-ci-operator-credentials",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "apici-ci-operator-credentials", Items: []kubeapi.KeyToPath{{Key: "sa.ci-operator.apici.config", Path: "kubeconfig"}}},
-						},
-					},
-					{
-						Name: "pull-secret",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "regcred"},
-						},
-					},
-					{
-						Name: "job-definition",
-						VolumeSource: kubeapi.VolumeSource{
-							ConfigMap: &kubeapi.ConfigMapVolumeSource{
-								LocalObjectReference: kubeapi.LocalObjectReference{
-									Name: "prow-job-cluster-launch-installer-custom-test-image",
-								},
-							},
-						},
-					},
-					{
-						Name: "cluster-profile",
-						VolumeSource: kubeapi.VolumeSource{
-							Projected: &kubeapi.ProjectedVolumeSource{
-								Sources: []kubeapi.VolumeProjection{
-									{
-										Secret: &kubeapi.SecretProjection{
-											LocalObjectReference: kubeapi.LocalObjectReference{
-												Name: "cluster-secrets-gcp",
-											},
-										},
-									},
-									{
-										ConfigMap: &kubeapi.ConfigMapProjection{
-											LocalObjectReference: kubeapi.LocalObjectReference{
-												Name: "cluster-profile-gcp",
-											},
-										},
-									},
-								},
-							},
-						},
-					},
-					{
-						Name: "boskos",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "boskos-credentials", Items: []kubeapi.KeyToPath{{Key: "password", Path: "password"}}},
-						},
-					},
-				},
-				Containers: []kubeapi.Container{{
-					Image:           "ci-operator:latest",
-					ImagePullPolicy: kubeapi.PullAlways,
-					Command:         []string{"ci-operator"},
-					Args: []string{
-						"--give-pr-author-access-to-namespace=true",
-						"--artifact-dir=$(ARTIFACTS)",
-						"--sentry-dsn-path=/etc/sentry-dsn/ci-operator",
-						"--kubeconfig=/etc/apici/kubeconfig",
-						"--image-import-pull-secret=/etc/pull-secret/.dockerconfigjson",
-						"--target=test",
-						"--secret-dir=/usr/local/test-cluster-profile",
-						"--template=/usr/local/test",
-						"--lease-server=https://boskos-ci.svc.ci.openshift.org",
-						"--lease-server-username=ci",
-						"--lease-server-password-file=/etc/boskos/password",
-					},
-					Resources: kubeapi.ResourceRequirements{
-						Requests: kubeapi.ResourceList{"cpu": *resource.NewMilliQuantity(10, resource.DecimalSI)},
-					},
-					Env: []kubeapi.EnvVar{
-						{Name: "CLUSTER_TYPE", Value: "gcp"},
-						{Name: "JOB_NAME_SAFE", Value: "test"},
-						{Name: "TEST_COMMAND", Value: "commands"},
-						{Name: "TEST_IMAGESTREAM_TAG", Value: "pipeline:kubevirt-test"},
-						{Name: "CLUSTER_ENABLE_NESTED_VIRT", Value: "true"},
-					},
-					VolumeMounts: []kubeapi.VolumeMount{
-						{Name: "sentry-dsn", MountPath: "/etc/sentry-dsn", ReadOnly: true},
-						{Name: "apici-ci-operator-credentials", ReadOnly: true, MountPath: "/etc/apici"},
-						{Name: "pull-secret", ReadOnly: true, MountPath: "/etc/pull-secret"},
-						{Name: "boskos", ReadOnly: true, MountPath: "/etc/boskos"},
-						{Name: "cluster-profile", MountPath: "/usr/local/test-cluster-profile"},
-						{Name: "job-definition", MountPath: "/usr/local/test", SubPath: "cluster-launch-installer-custom-test-image.yaml"},
-					},
-				}},
-			},
-		},
-		{
-			info:    &prowgenInfo{Info: config.Info{Org: "organization", Repo: "repo", Branch: "branch"}},
-			release: "origin-v4.0",
-			test: ciop.TestStepConfiguration{
-				As:       "test",
-				Commands: "commands",
-				OpenshiftInstallerCustomTestImageClusterTestConfiguration: &ciop.OpenshiftInstallerCustomTestImageClusterTestConfiguration{
-					ClusterTestConfiguration: ciop.ClusterTestConfiguration{ClusterProfile: "gcp"},
-					From:                     "pipeline:kubevirt-test",
-					NestedVirtImage:          "",
-					EnableNestedVirt:         false,
-				},
-			},
-
-			expected: &kubeapi.PodSpec{
-				ServiceAccountName: "ci-operator",
-				Volumes: []kubeapi.Volume{
-					{
-						Name: "sentry-dsn",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "sentry-dsn"},
-						},
-					},
-					{
-						Name: "apici-ci-operator-credentials",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "apici-ci-operator-credentials", Items: []kubeapi.KeyToPath{{Key: "sa.ci-operator.apici.config", Path: "kubeconfig"}}},
-						},
-					},
-					{
-						Name: "pull-secret",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "regcred"},
-						},
-					},
-					{
-						Name: "job-definition",
-						VolumeSource: kubeapi.VolumeSource{
-							ConfigMap: &kubeapi.ConfigMapVolumeSource{
-								LocalObjectReference: kubeapi.LocalObjectReference{
-									Name: "prow-job-cluster-launch-installer-custom-test-image",
-								},
-							},
-						},
-					},
-					{
-						Name: "cluster-profile",
-						VolumeSource: kubeapi.VolumeSource{
-							Projected: &kubeapi.ProjectedVolumeSource{
-								Sources: []kubeapi.VolumeProjection{
-									{
-										Secret: &kubeapi.SecretProjection{
-											LocalObjectReference: kubeapi.LocalObjectReference{
-												Name: "cluster-secrets-gcp",
-											},
-										},
-									},
-									{
-										ConfigMap: &kubeapi.ConfigMapProjection{
-											LocalObjectReference: kubeapi.LocalObjectReference{
-												Name: "cluster-profile-gcp",
-											},
-										},
-									},
-								},
-							},
-						},
-					},
-					{
-						Name: "boskos",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "boskos-credentials", Items: []kubeapi.KeyToPath{{Key: "password", Path: "password"}}},
-						},
-					},
-				},
-				Containers: []kubeapi.Container{{
-					Image:           "ci-operator:latest",
-					ImagePullPolicy: kubeapi.PullAlways,
-					Command:         []string{"ci-operator"},
-					Args: []string{
-						"--give-pr-author-access-to-namespace=true",
-						"--artifact-dir=$(ARTIFACTS)",
-						"--sentry-dsn-path=/etc/sentry-dsn/ci-operator",
-						"--kubeconfig=/etc/apici/kubeconfig",
-						"--image-import-pull-secret=/etc/pull-secret/.dockerconfigjson",
-						"--target=test",
-						"--secret-dir=/usr/local/test-cluster-profile",
-						"--template=/usr/local/test",
-						"--lease-server=https://boskos-ci.svc.ci.openshift.org",
-						"--lease-server-username=ci",
-						"--lease-server-password-file=/etc/boskos/password",
-					},
-					Resources: kubeapi.ResourceRequirements{
-						Requests: kubeapi.ResourceList{"cpu": *resource.NewMilliQuantity(10, resource.DecimalSI)},
-					},
-					Env: []kubeapi.EnvVar{
-						{Name: "CLUSTER_TYPE", Value: "gcp"},
-						{Name: "JOB_NAME_SAFE", Value: "test"},
-						{Name: "TEST_COMMAND", Value: "commands"},
-						{Name: "TEST_IMAGESTREAM_TAG", Value: "pipeline:kubevirt-test"},
-					},
-					VolumeMounts: []kubeapi.VolumeMount{
-						{Name: "sentry-dsn", MountPath: "/etc/sentry-dsn", ReadOnly: true},
-						{Name: "apici-ci-operator-credentials", ReadOnly: true, MountPath: "/etc/apici"},
-						{Name: "pull-secret", ReadOnly: true, MountPath: "/etc/pull-secret"},
-						{Name: "boskos", ReadOnly: true, MountPath: "/etc/boskos"},
-						{Name: "cluster-profile", MountPath: "/usr/local/test-cluster-profile"},
-						{Name: "job-definition", MountPath: "/usr/local/test", SubPath: "cluster-launch-installer-custom-test-image.yaml"},
-					},
-				}},
-			},
-		},
-		{
-			info:    &prowgenInfo{Info: config.Info{Org: "organization", Repo: "repo", Branch: "branch"}},
-			release: "origin-v4.0",
-			test: ciop.TestStepConfiguration{
-				As:       "test",
-				Commands: "commands",
-				OpenshiftInstallerCustomTestImageClusterTestConfiguration: &ciop.OpenshiftInstallerCustomTestImageClusterTestConfiguration{
-					ClusterTestConfiguration: ciop.ClusterTestConfiguration{ClusterProfile: "gcp"},
-					From:                     "pipeline:kubevirt-test",
-				},
-			},
-
-			expected: &kubeapi.PodSpec{
-				ServiceAccountName: "ci-operator",
-				Volumes: []kubeapi.Volume{
-					{
-						Name: "sentry-dsn",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "sentry-dsn"},
-						},
-					},
-					{
-						Name: "apici-ci-operator-credentials",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "apici-ci-operator-credentials", Items: []kubeapi.KeyToPath{{Key: "sa.ci-operator.apici.config", Path: "kubeconfig"}}},
-						},
-					},
-					{
-						Name: "pull-secret",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "regcred"},
-						},
-					},
-					{
-						Name: "job-definition",
-						VolumeSource: kubeapi.VolumeSource{
-							ConfigMap: &kubeapi.ConfigMapVolumeSource{
-								LocalObjectReference: kubeapi.LocalObjectReference{
-									Name: "prow-job-cluster-launch-installer-custom-test-image",
-								},
-							},
-						},
-					},
-					{
-						Name: "cluster-profile",
-						VolumeSource: kubeapi.VolumeSource{
-							Projected: &kubeapi.ProjectedVolumeSource{
-								Sources: []kubeapi.VolumeProjection{
-									{
-										Secret: &kubeapi.SecretProjection{
-											LocalObjectReference: kubeapi.LocalObjectReference{
-												Name: "cluster-secrets-gcp",
-											},
-										},
-									},
-									{
-										ConfigMap: &kubeapi.ConfigMapProjection{
-											LocalObjectReference: kubeapi.LocalObjectReference{
-												Name: "cluster-profile-gcp",
-											},
-										},
-									},
-								},
-							},
-						},
-					},
-					{
-						Name: "boskos",
-						VolumeSource: kubeapi.VolumeSource{
-							Secret: &kubeapi.SecretVolumeSource{SecretName: "boskos-credentials", Items: []kubeapi.KeyToPath{{Key: "password", Path: "password"}}},
-						},
-					},
-				},
-				Containers: []kubeapi.Container{{
-					Image:           "ci-operator:latest",
-					ImagePullPolicy: kubeapi.PullAlways,
-					Command:         []string{"ci-operator"},
-					Args: []string{
-						"--give-pr-author-access-to-namespace=true",
-						"--artifact-dir=$(ARTIFACTS)",
-						"--sentry-dsn-path=/etc/sentry-dsn/ci-operator",
-						"--kubeconfig=/etc/apici/kubeconfig",
-						"--image-import-pull-secret=/etc/pull-secret/.dockerconfigjson",
-						"--target=test",
-						"--secret-dir=/usr/local/test-cluster-profile",
-						"--template=/usr/local/test",
-						"--lease-server=https://boskos-ci.svc.ci.openshift.org",
-						"--lease-server-username=ci",
-						"--lease-server-password-file=/etc/boskos/password",
-					},
-					Resources: kubeapi.ResourceRequirements{
-						Requests: kubeapi.ResourceList{"cpu": *resource.NewMilliQuantity(10, resource.DecimalSI)},
-					},
-					Env: []kubeapi.EnvVar{
-						{Name: "CLUSTER_TYPE", Value: "gcp"},
-						{Name: "JOB_NAME_SAFE", Value: "test"},
-						{Name: "TEST_COMMAND", Value: "commands"},
-						{Name: "TEST_IMAGESTREAM_TAG", Value: "pipeline:kubevirt-test"},
-					},
-					VolumeMounts: []kubeapi.VolumeMount{
-						{Name: "sentry-dsn", MountPath: "/etc/sentry-dsn", ReadOnly: true},
-						{Name: "apici-ci-operator-credentials", ReadOnly: true, MountPath: "/etc/apici"},
-						{Name: "pull-secret", ReadOnly: true, MountPath: "/etc/pull-secret"},
-						{Name: "boskos", ReadOnly: true, MountPath: "/etc/boskos"},
-						{Name: "cluster-profile", MountPath: "/usr/local/test-cluster-profile"},
-						{Name: "job-definition", MountPath: "/usr/local/test", SubPath: "cluster-launch-installer-custom-test-image.yaml"},
-					},
-				}},
-			},
-		},
-	}
-
-	for _, tc := range tests {
-		podSpec := generatePodSpecOthers(tc.info, tc.release, &tc.test)
-		if !equality.Semantic.DeepEqual(podSpec, tc.expected) {
-			t.Errorf("expected PodSpec diff:\n%s", cmp.Diff(tc.expected, podSpec, unexportedFields...))
-		}
-	}
-}
-
-func TestGeneratePresubmitForTest(t *testing.T) {
-	newTrue := true
-
-	tests := []struct {
-		description string
-
-		test     string
-		repoInfo *prowgenInfo
-		expected *prowconfig.Presubmit
-	}{{
-		description: "presubmit for standard test",
-		test:        "testname",
-		repoInfo:    &prowgenInfo{Info: config.Info{Org: "org", Repo: "repo", Branch: "branch"}},
-
-		expected: &prowconfig.Presubmit{
-			JobBase: prowconfig.JobBase{
-				Agent: "kubernetes",
-				Labels: map[string]string{
-					"ci-operator.openshift.io/prowgen-controlled": "true",
-					"pj-rehearse.openshift.io/can-be-rehearsed":   "true",
-				},
-				Name: "pull-ci-org-repo-branch-testname",
-				UtilityConfig: prowconfig.UtilityConfig{
-					DecorationConfig: &v1.DecorationConfig{SkipCloning: &newTrue},
-					Decorate:         true,
-				},
-			},
-			AlwaysRun: true,
-			Brancher:  prowconfig.Brancher{Branches: []string{"branch"}},
-			Reporter: prowconfig.Reporter{
-				Context: "ci/prow/testname",
-			},
-			RerunCommand: "/test testname",
-			Trigger:      `(?m)^/test( | .* )testname,?($|\s.*)`,
-		},
-	},
-		{
-			description: "presubmit for a test in a variant config",
-			test:        "testname",
-			repoInfo:    &prowgenInfo{Info: config.Info{Org: "org", Repo: "repo", Branch: "branch", Variant: "also"}},
-
-			expected: &prowconfig.Presubmit{
-				JobBase: prowconfig.JobBase{
-					Agent: "kubernetes",
-					Labels: map[string]string{
-						"ci-operator.openshift.io/prowgen-controlled": "true",
-						"ci-operator.openshift.io/variant":            "also",
-						"pj-rehearse.openshift.io/can-be-rehearsed":   "true",
-					},
-					Name: "pull-ci-org-repo-branch-also-testname",
-					UtilityConfig: prowconfig.UtilityConfig{
-						DecorationConfig: &v1.DecorationConfig{SkipCloning: &newTrue},
-						Decorate:         true,
-					},
-				},
-				AlwaysRun: true,
-				Brancher:  prowconfig.Brancher{Branches: []string{"branch"}},
-				Reporter: prowconfig.Reporter{
-					Context: "ci/prow/also-testname",
-				},
-				RerunCommand: "/test also-testname",
-				Trigger:      `(?m)^/test( | .* )also-testname,?($|\s.*)`,
-			},
-		},
-	}
-	for _, tc := range tests {
-		presubmit := generatePresubmitForTest(tc.test, tc.repoInfo, jobconfig.Generated, nil, true, nil) // podSpec tested in generatePodSpec
-		if !equality.Semantic.DeepEqual(presubmit, tc.expected) {
-			t.Errorf("expected presubmit diff:\n%s", cmp.Diff(tc.expected, presubmit, unexportedFields...))
-		}
-	}
-}
-
-func TestGeneratePeriodicForTest(t *testing.T) {
-	newTrue := true
-
-	tests := []struct {
-		description string
-
-		test     string
-		repoInfo *prowgenInfo
-		expected *prowconfig.Periodic
-	}{{
-		description: "periodic for standard test",
-		test:        "testname",
-		repoInfo:    &prowgenInfo{Info: config.Info{Org: "org", Repo: "repo", Branch: "branch"}},
-
-		expected: &prowconfig.Periodic{
-			Cron: "@yearly",
-			JobBase: prowconfig.JobBase{
-				Agent: "kubernetes",
-				Labels: map[string]string{
-					"ci-operator.openshift.io/prowgen-controlled": "true",
-					"pj-rehearse.openshift.io/can-be-rehearsed":   "true",
-				},
-				Name: "periodic-ci-org-repo-branch-testname",
-				UtilityConfig: prowconfig.UtilityConfig{
-					DecorationConfig: &v1.DecorationConfig{SkipCloning: &newTrue},
-					Decorate:         true,
-					ExtraRefs: []v1.Refs{{
-						Org:     "org",
-						Repo:    "repo",
-						BaseRef: "branch",
-					}},
-				},
-			},
-		},
-	},
-		{
-			description: "periodic for a test in a variant config",
-			test:        "testname",
-			repoInfo:    &prowgenInfo{Info: config.Info{Org: "org", Repo: "repo", Branch: "branch", Variant: "also"}},
-			expected: &prowconfig.Periodic{
-				Cron: "@yearly",
-				JobBase: prowconfig.JobBase{
-					Agent: "kubernetes",
-					Labels: map[string]string{
-						"ci-operator.openshift.io/prowgen-controlled": "true",
-						"ci-operator.openshift.io/variant":            "also",
-						"pj-rehearse.openshift.io/can-be-rehearsed":   "true",
-					},
-					Name: "periodic-ci-org-repo-branch-also-testname",
-					UtilityConfig: prowconfig.UtilityConfig{
-						DecorationConfig: &v1.DecorationConfig{SkipCloning: &newTrue},
-						Decorate:         true,
-						ExtraRefs: []v1.Refs{{
-							Org:     "org",
-							Repo:    "repo",
-							BaseRef: "branch",
-						}},
-					},
-				},
-			},
-		},
-	}
-	for _, tc := range tests {
-		periodic := generatePeriodicForTest(tc.test, tc.repoInfo, jobconfig.Generated, nil, true, "@yearly", nil) // podSpec tested in generatePodSpec
-		// Periodic has unexported fields on which DeepEqual panics, so we need to compare member-wise
-		if !equality.Semantic.DeepEqual(periodic.JobBase, tc.expected.JobBase) {
-			t.Errorf("expected periodic diff:\n%s", cmp.Diff(tc.expected.JobBase, periodic.JobBase, unexportedFields...))
-		}
-		if periodic.Cron != tc.expected.Cron {
-			t.Errorf("expected periodic cron diff:\n%s", cmp.Diff(tc.expected.Cron, periodic.Cron, unexportedFields...))
-		}
-	}
-}
-
-func TestGeneratePostSubmitForTest(t *testing.T) {
-	newTrue := true
-	standardJobLabels := map[string]string{"ci-operator.openshift.io/prowgen-controlled": "true"}
-	tests := []struct {
-		name     string
-		repoInfo *prowgenInfo
-
-		expected *prowconfig.Postsubmit
-	}{
-		{
-			name: "name",
-			repoInfo: &prowgenInfo{Info: config.Info{
-				Org:    "organization",
-				Repo:   "repository",
-				Branch: "branch",
-			}},
-
-			expected: &prowconfig.Postsubmit{
-				JobBase: prowconfig.JobBase{
-					Agent:  "kubernetes",
-					Labels: standardJobLabels,
-					Name:   "branch-ci-organization-repository-branch-name",
-					UtilityConfig: prowconfig.UtilityConfig{
-						DecorationConfig: &v1.DecorationConfig{SkipCloning: &newTrue},
-						Decorate:         true,
-					},
-				},
-
-				Brancher: prowconfig.Brancher{Branches: []string{"^branch$"}},
-			},
-		},
-		{
-			name: "Name",
-			repoInfo: &prowgenInfo{Info: config.Info{
-				Org:    "Organization",
-				Repo:   "Repository",
-				Branch: "Branch",
-			}},
-
-			expected: &prowconfig.Postsubmit{
-				JobBase: prowconfig.JobBase{
-					Agent:  "kubernetes",
-					Name:   "branch-ci-Organization-Repository-Branch-Name",
-					Labels: map[string]string{"ci-operator.openshift.io/prowgen-controlled": "true"},
-					UtilityConfig: prowconfig.UtilityConfig{
-						DecorationConfig: &v1.DecorationConfig{SkipCloning: &newTrue},
-						Decorate:         true,
-					}},
-				Brancher: prowconfig.Brancher{Branches: []string{"^Branch$"}},
-			},
-		},
-		{
-			name: "name",
-			repoInfo: &prowgenInfo{Info: config.Info{
-				Org:    "Organization",
-				Repo:   "Repository",
-				Branch: "Branch",
-			}},
-
-			expected: &prowconfig.Postsubmit{
-				JobBase: prowconfig.JobBase{
-					Agent:  "kubernetes",
-					Name:   "branch-ci-Organization-Repository-Branch-name",
-					Labels: map[string]string{"ci-operator.openshift.io/prowgen-controlled": "true"},
-					UtilityConfig: prowconfig.UtilityConfig{
-						DecorationConfig: &v1.DecorationConfig{SkipCloning: &newTrue},
-						Decorate:         true,
-					}},
-				Brancher: prowconfig.Brancher{Branches: []string{"^Branch$"}},
-			},
-		},
-	}
-	for _, tc := range tests {
-		postsubmit := generatePostsubmitForTest(tc.name, tc.repoInfo, jobconfig.Generated, nil, nil) // podSpec tested in TestGeneratePodSpec
-		if !equality.Semantic.DeepEqual(postsubmit, tc.expected) {
-			t.Errorf("expected postsubmit diff:\n%s", cmp.Diff(tc.expected, postsubmit, unexportedFields...))
-		}
-	}
-}
-
-func TestGenerateJobs(t *testing.T) {
-	standardPresubmitJobLabels := map[string]string{
-		"ci-operator.openshift.io/prowgen-controlled": "true",
-		"pj-rehearse.openshift.io/can-be-rehearsed":   "true"}
-	standardPostsubmitJobLabels := map[string]string{"ci-operator.openshift.io/prowgen-controlled": "true"}
-
-	tests := []struct {
-		id       string
-		config   *ciop.ReleaseBuildConfiguration
-		repoInfo *prowgenInfo
-		expected *prowconfig.JobConfig
-	}{
-		{
-			id: "two tests and empty Images so only two test presubmits are generated",
-			config: &ciop.ReleaseBuildConfiguration{
-				Tests: []ciop.TestStepConfiguration{
-					{As: "derTest", ContainerTestConfiguration: &ciop.ContainerTestConfiguration{From: "from"}},
-					{As: "leTest", ContainerTestConfiguration: &ciop.ContainerTestConfiguration{From: "from"}}},
-			},
-			repoInfo: &prowgenInfo{Info: config.Info{
-				Org:    "organization",
-				Repo:   "repository",
-				Branch: "branch",
-			}},
-			expected: &prowconfig.JobConfig{
-				PresubmitsStatic: map[string][]prowconfig.Presubmit{"organization/repository": {{
-					JobBase: prowconfig.JobBase{
-						Name:   "pull-ci-organization-repository-branch-derTest",
-						Labels: standardPresubmitJobLabels,
-					}}, {
-					JobBase: prowconfig.JobBase{
-						Name:   "pull-ci-organization-repository-branch-leTest",
-						Labels: standardPresubmitJobLabels,
-					}},
-				}},
-				PostsubmitsStatic: map[string][]prowconfig.Postsubmit{},
-			},
-		}, {
-			id: "two tests and nonempty Images so two test presubmits and images pre/postsubmits are generated ",
-			config: &ciop.ReleaseBuildConfiguration{
-				Tests: []ciop.TestStepConfiguration{
-					{As: "derTest", ContainerTestConfiguration: &ciop.ContainerTestConfiguration{From: "from"}},
-					{As: "leTest", ContainerTestConfiguration: &ciop.ContainerTestConfiguration{From: "from"}}},
-				Images:                 []ciop.ProjectDirectoryImageBuildStepConfiguration{{}},
-				PromotionConfiguration: &ciop.PromotionConfiguration{},
-			},
-			repoInfo: &prowgenInfo{Info: config.Info{
-				Org:    "organization",
-				Repo:   "repository",
-				Branch: "branch",
-			}},
-			expected: &prowconfig.JobConfig{
-				PresubmitsStatic: map[string][]prowconfig.Presubmit{"organization/repository": {{
-					JobBase: prowconfig.JobBase{
-						Name:   "pull-ci-organization-repository-branch-derTest",
-						Labels: standardPresubmitJobLabels,
-					}}, {
-					JobBase: prowconfig.JobBase{
-						Name:   "pull-ci-organization-repository-branch-leTest",
-						Labels: standardPresubmitJobLabels,
-					}}, {
-					JobBase: prowconfig.JobBase{
-						Name:   "pull-ci-organization-repository-branch-images",
-						Labels: standardPresubmitJobLabels,
-					}},
-				}},
-				PostsubmitsStatic: map[string][]prowconfig.Postsubmit{"organization/repository": {{
-					JobBase: prowconfig.JobBase{
-						Name:           "branch-ci-organization-repository-branch-images",
-						Labels:         standardPostsubmitJobLabels,
-						MaxConcurrency: 1,
-					}},
-				}},
-			},
-		}, {
-			id: "template test",
-			config: &ciop.ReleaseBuildConfiguration{
-				InputConfiguration: ciop.InputConfiguration{
-					ReleaseTagConfiguration: &ciop.ReleaseTagConfiguration{Name: "origin-v4.0"}},
-				Tests: []ciop.TestStepConfiguration{
-					{
-						As: "oTeste",
-						OpenshiftAnsibleClusterTestConfiguration: &ciop.OpenshiftAnsibleClusterTestConfiguration{
-							ClusterTestConfiguration: ciop.ClusterTestConfiguration{ClusterProfile: "gcp"},
-						},
-					},
-				},
-			},
-			repoInfo: &prowgenInfo{Info: config.Info{
-				Org:    "organization",
-				Repo:   "repository",
-				Branch: "branch",
-			}},
-			expected: &prowconfig.JobConfig{
-				PresubmitsStatic: map[string][]prowconfig.Presubmit{"organization/repository": {{
-					JobBase: prowconfig.JobBase{
-						Name:   "pull-ci-organization-repository-branch-oTeste",
-						Labels: standardPresubmitJobLabels,
-					}},
-				}},
-			},
-		}, {
-			id: "template test which doesn't require `tag_specification`",
-			config: &ciop.ReleaseBuildConfiguration{
-				Tests: []ciop.TestStepConfiguration{{
-					As: "oTeste",
-					OpenshiftInstallerClusterTestConfiguration: &ciop.OpenshiftInstallerClusterTestConfiguration{
-						ClusterTestConfiguration: ciop.ClusterTestConfiguration{ClusterProfile: "gcp"},
-					},
-				}},
-			},
-			repoInfo: &prowgenInfo{Info: config.Info{
-				Org:    "organization",
-				Repo:   "repository",
-				Branch: "branch",
-			}},
-			expected: &prowconfig.JobConfig{
-				PresubmitsStatic: map[string][]prowconfig.Presubmit{"organization/repository": {{
-					JobBase: prowconfig.JobBase{
-						Name:   "pull-ci-organization-repository-branch-oTeste",
-						Labels: standardPresubmitJobLabels,
-					}},
-				}},
-			},
-		}, {
-			id: "Promotion configuration causes --promote job",
-			config: &ciop.ReleaseBuildConfiguration{
-				Tests:                  []ciop.TestStepConfiguration{},
-				Images:                 []ciop.ProjectDirectoryImageBuildStepConfiguration{{}},
-				PromotionConfiguration: &ciop.PromotionConfiguration{Namespace: "ci"},
-			},
-			repoInfo: &prowgenInfo{Info: config.Info{
-				Org:    "organization",
-				Repo:   "repository",
-				Branch: "branch",
-			}},
-			expected: &prowconfig.JobConfig{
-				PresubmitsStatic: map[string][]prowconfig.Presubmit{"organization/repository": {{
-					JobBase: prowconfig.JobBase{
-						Name:   "pull-ci-organization-repository-branch-images",
-						Labels: standardPresubmitJobLabels,
-					}},
-				}},
-				PostsubmitsStatic: map[string][]prowconfig.Postsubmit{"organization/repository": {{
-					JobBase: prowconfig.JobBase{
-						Name:           "branch-ci-organization-repository-branch-images",
-						Labels:         standardPostsubmitJobLabels,
-						MaxConcurrency: 1,
-					}},
-				}},
-			},
-		}, {
-			id: "no Promotion configuration has no branch job",
-			config: &ciop.ReleaseBuildConfiguration{
-				Tests:  []ciop.TestStepConfiguration{},
-				Images: []ciop.ProjectDirectoryImageBuildStepConfiguration{{}},
-				InputConfiguration: ciop.InputConfiguration{
-					ReleaseTagConfiguration: &ciop.ReleaseTagConfiguration{Namespace: "openshift"},
-				},
-			},
-			repoInfo: &prowgenInfo{Info: config.Info{
-				Org:    "organization",
-				Repo:   "repository",
-				Branch: "branch",
-			}},
-			expected: &prowconfig.JobConfig{
-				PresubmitsStatic: map[string][]prowconfig.Presubmit{"organization/repository": {{
-					JobBase: prowconfig.JobBase{
-						Name:   "pull-ci-organization-repository-branch-images",
-						Labels: standardPresubmitJobLabels,
-					}},
-				}},
-			},
-		},
-	}
-
-	log.SetOutput(ioutil.Discard)
-	for _, tc := range tests {
-		jobConfig := generateJobs(tc.config, tc.repoInfo, jobconfig.Generated)
-
-		pruneForTests(jobConfig) // prune the fields that are tested in TestGeneratePre/PostsubmitForTest
-
-		if !equality.Semantic.DeepEqual(jobConfig, tc.expected) {
-			t.Errorf("testcase: %s\nexpected job config diff:\n%s", tc.id, cmp.Diff(tc.expected, jobConfig, unexportedFields...))
-		}
-	}
-}
-
-func pruneForTests(jobConfig *prowconfig.JobConfig) {
-	for repo := range jobConfig.PresubmitsStatic {
-		for i := range jobConfig.PresubmitsStatic[repo] {
-			jobConfig.PresubmitsStatic[repo][i].AlwaysRun = false
-			jobConfig.PresubmitsStatic[repo][i].Context = ""
-			jobConfig.PresubmitsStatic[repo][i].Trigger = ""
-			jobConfig.PresubmitsStatic[repo][i].RerunCommand = ""
-			jobConfig.PresubmitsStatic[repo][i].Agent = ""
-			jobConfig.PresubmitsStatic[repo][i].Spec = nil
-			jobConfig.PresubmitsStatic[repo][i].Brancher = prowconfig.Brancher{}
-			jobConfig.PresubmitsStatic[repo][i].UtilityConfig = prowconfig.UtilityConfig{}
-		}
-	}
-	for repo := range jobConfig.PostsubmitsStatic {
-		for i := range jobConfig.PostsubmitsStatic[repo] {
-			jobConfig.PostsubmitsStatic[repo][i].Agent = ""
-			jobConfig.PostsubmitsStatic[repo][i].Spec = nil
-			jobConfig.PostsubmitsStatic[repo][i].Brancher = prowconfig.Brancher{}
-			jobConfig.PostsubmitsStatic[repo][i].UtilityConfig = prowconfig.UtilityConfig{}
-		}
-	}
 }
 
 func TestFromCIOperatorConfigToProwYaml(t *testing.T) {
@@ -1840,166 +347,6 @@ func TestPruneStaleJobs(t *testing.T) {
 
 			if pruned := prune(tc.jobconfig); !reflect.DeepEqual(pruned, expected) {
 				t.Errorf("Pruned config differs:\n%s", cmp.Diff(expected, pruned, unexportedFields...))
-			}
-		})
-	}
-}
-
-func TestGenerateJobBase(t *testing.T) {
-	yes := true
-	path := "/some/where"
-	var testCases = []struct {
-		testName    string
-		name        string
-		prefix      string
-		info        *prowgenInfo
-		label       jobconfig.ProwgenLabel
-		podSpec     *kubeapi.PodSpec
-		rehearsable bool
-		pathAlias   *string
-		expected    prowconfig.JobBase
-	}{
-		{
-			testName: "no special options",
-			name:     "test",
-			prefix:   "pull",
-			info:     &prowgenInfo{Info: config.Info{Org: "org", Repo: "repo", Branch: "branch"}},
-			label:    jobconfig.Generated,
-			podSpec:  &kubeapi.PodSpec{Containers: []kubeapi.Container{{Name: "test"}}},
-			expected: prowconfig.JobBase{
-				Name:  "pull-ci-org-repo-branch-test",
-				Agent: "kubernetes",
-				Labels: map[string]string{
-					"ci-operator.openshift.io/prowgen-controlled": "true",
-				},
-				Spec:          &kubeapi.PodSpec{Containers: []kubeapi.Container{{Name: "test"}}},
-				UtilityConfig: prowconfig.UtilityConfig{Decorate: true, DecorationConfig: &v1.DecorationConfig{SkipCloning: &yes}},
-			},
-		},
-		{
-			testName:    "rehearsable",
-			name:        "test",
-			prefix:      "pull",
-			info:        &prowgenInfo{Info: config.Info{Org: "org", Repo: "repo", Branch: "branch"}},
-			label:       jobconfig.Generated,
-			podSpec:     &kubeapi.PodSpec{Containers: []kubeapi.Container{{Name: "test"}}},
-			rehearsable: true,
-			expected: prowconfig.JobBase{
-				Name:  "pull-ci-org-repo-branch-test",
-				Agent: "kubernetes",
-				Labels: map[string]string{
-					"ci-operator.openshift.io/prowgen-controlled": "true",
-					"pj-rehearse.openshift.io/can-be-rehearsed":   "true",
-				},
-				Spec:          &kubeapi.PodSpec{Containers: []kubeapi.Container{{Name: "test"}}},
-				UtilityConfig: prowconfig.UtilityConfig{Decorate: true, DecorationConfig: &v1.DecorationConfig{SkipCloning: &yes}},
-			},
-		},
-		{
-			testName: "config variant",
-			name:     "test",
-			prefix:   "pull",
-			info:     &prowgenInfo{Info: config.Info{Org: "org", Repo: "repo", Branch: "branch", Variant: "whatever"}},
-			label:    jobconfig.Generated,
-			podSpec:  &kubeapi.PodSpec{Containers: []kubeapi.Container{{Name: "test"}}},
-			expected: prowconfig.JobBase{
-				Name:  "pull-ci-org-repo-branch-whatever-test",
-				Agent: "kubernetes",
-				Labels: map[string]string{
-					"ci-operator.openshift.io/prowgen-controlled": "true",
-					"ci-operator.openshift.io/variant":            "whatever",
-				},
-				Spec:          &kubeapi.PodSpec{Containers: []kubeapi.Container{{Name: "test"}}},
-				UtilityConfig: prowconfig.UtilityConfig{Decorate: true, DecorationConfig: &v1.DecorationConfig{SkipCloning: &yes}},
-			},
-		},
-		{
-			testName:  "path alias",
-			name:      "test",
-			prefix:    "pull",
-			info:      &prowgenInfo{Info: config.Info{Org: "org", Repo: "repo", Branch: "branch", Variant: "whatever"}},
-			label:     jobconfig.Generated,
-			podSpec:   &kubeapi.PodSpec{Containers: []kubeapi.Container{{Name: "test"}}},
-			pathAlias: &path,
-			expected: prowconfig.JobBase{
-				Name:  "pull-ci-org-repo-branch-whatever-test",
-				Agent: "kubernetes",
-				Labels: map[string]string{
-					"ci-operator.openshift.io/prowgen-controlled": "true",
-					"ci-operator.openshift.io/variant":            "whatever",
-				},
-				Spec:          &kubeapi.PodSpec{Containers: []kubeapi.Container{{Name: "test"}}},
-				UtilityConfig: prowconfig.UtilityConfig{Decorate: true, DecorationConfig: &v1.DecorationConfig{SkipCloning: &yes}, PathAlias: "/some/where"},
-			},
-		},
-		{
-			testName: "hidden job for private repos",
-			name:     "test",
-			prefix:   "pull",
-			info: &prowgenInfo{
-				Info:   config.Info{Org: "org", Repo: "repo", Branch: "branch"},
-				config: config.Prowgen{Private: true},
-			},
-			label:   jobconfig.Generated,
-			podSpec: &kubeapi.PodSpec{Containers: []kubeapi.Container{{Name: "test"}}},
-			expected: prowconfig.JobBase{
-				Name:  "pull-ci-org-repo-branch-test",
-				Agent: "kubernetes",
-				Labels: map[string]string{
-					"ci-operator.openshift.io/prowgen-controlled": "true",
-				},
-				Spec:          &kubeapi.PodSpec{Containers: []kubeapi.Container{{Name: "test"}}},
-				UtilityConfig: prowconfig.UtilityConfig{Decorate: true, DecorationConfig: &v1.DecorationConfig{SkipCloning: &yes}},
-				Hidden:        true,
-			},
-		},
-		{
-			testName: "expose job for private repos with public results",
-			name:     "test",
-			prefix:   "pull",
-			info: &prowgenInfo{
-				Info:   config.Info{Org: "org", Repo: "repo", Branch: "branch"},
-				config: config.Prowgen{Private: true, Expose: true},
-			},
-			label:   jobconfig.Generated,
-			podSpec: &kubeapi.PodSpec{Containers: []kubeapi.Container{{Name: "test"}}},
-			expected: prowconfig.JobBase{
-				Name:  "pull-ci-org-repo-branch-test",
-				Agent: "kubernetes",
-				Labels: map[string]string{
-					"ci-operator.openshift.io/prowgen-controlled": "true",
-				},
-				Spec:          &kubeapi.PodSpec{Containers: []kubeapi.Container{{Name: "test"}}},
-				UtilityConfig: prowconfig.UtilityConfig{Decorate: true, DecorationConfig: &v1.DecorationConfig{SkipCloning: &yes}},
-				Hidden:        false,
-			},
-		},
-		{
-			testName: "expose option set but not private",
-			name:     "test",
-			prefix:   "pull",
-			info: &prowgenInfo{
-				Info:   config.Info{Org: "org", Repo: "repo", Branch: "branch"},
-				config: config.Prowgen{Private: false, Expose: true},
-			},
-			label:   jobconfig.Generated,
-			podSpec: &kubeapi.PodSpec{Containers: []kubeapi.Container{{Name: "test"}}},
-			expected: prowconfig.JobBase{
-				Name:  "pull-ci-org-repo-branch-test",
-				Agent: "kubernetes",
-				Labels: map[string]string{
-					"ci-operator.openshift.io/prowgen-controlled": "true",
-				},
-				Spec:          &kubeapi.PodSpec{Containers: []kubeapi.Container{{Name: "test"}}},
-				UtilityConfig: prowconfig.UtilityConfig{Decorate: true, DecorationConfig: &v1.DecorationConfig{SkipCloning: &yes}},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.testName, func(t *testing.T) {
-			if actual, expected := generateJobBase(testCase.name, testCase.prefix, testCase.info, testCase.label, testCase.podSpec, testCase.rehearsable, testCase.pathAlias), testCase.expected; !reflect.DeepEqual(actual, expected) {
-				t.Errorf("%s: got incorrect job base: %v", testCase.testName, cmp.Diff(actual, expected, unexportedFields...))
 			}
 		})
 	}
